requirements.txt
fastapi
uvicorn
pandas
python-dotenv
openai
langchain
langchain-openai
langchain-anthropic
pydantic
requests
langchain-groq